requirements.txt
streamlit
cohere
pinecone-client
langchain-openai>=0.3.32
langchain-core>=0.3.75
langgraph>=0.6.6
python-dotenv==1.0.0
pydantic>=2.9.0,<3.0
